pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sequence-former"
version = "0.1.2"
authors = [
  { name="IndenScale", email="hfu7565ytru@gmail.com" },
]
description = "An intelligent, multi-modal document structuring engine powered by LLMs."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Text Processing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.0",
    "openai>=1.0",
    "python-dotenv>=1.0",
]

[project.urls]
"Homepage" = "https://github.com/IndenScale/SequenceFormer"
"Bug Tracker" = "https://github.com/IndenScale/SequenceFormer/issues"

[project.scripts]
sequence-former = "sequence_former.main:main"

[tool.setuptools.packages.find]
where = ["src"]
